requirements.txt
streamlit~=1.29.0
openai~=1.3.7
ipython~=8.18.1
langchain~=0.0.345
python-dotenv~=1.0.0